+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2020-12-11
+### Added
+- Possibility to specify tag name and tag values for getting subnets and security groups. 
+Also, support for security groups names.  
+### Changed
+- ***Important!*** The `subnetNames` and `securityGroupNames` options have been deprecated but it still will work for a while. 
+The new option are `subnets` and `securityGroups`.
+
 ## [2.2.1] - 2020-12-02
 ### Changed
 - Fixed travis build